Fix decide endpoint
